fix: fp16-safe additive masks; bf16 e2e stable; README precision guidance
Mask bias in triangle attention and MSA pair-weighted-averaging now built in fp32 so the large inf constant does not saturate to +inf under fp16 (fully-masked softmax rows were going NaN). fp32 path bit-identical (88 passed). bf16+XLA e2e stable (~0.26A aligned RMSD vs fp32); fp16 documented as range-unstable. README/known-limitations updated.

+++ src/boltz_jax/models/triangle/triangle_attention.py
@@ -101,7 +101,10 @@ def triangle_attention_forward(
         eps,
     )
     mask = mask[..., :, None, None, :]
-    mask_bias = inf * (mask - 1.0)
+    # Build the additive mask in fp32 so the large `inf` constant stays finite:
+    # in fp16, `inf` (1e9) would saturate to +inf and a fully-masked softmax row
+    # would become 0/0 -> NaN. fp32 runtime is unchanged (mask already fp32).
+    mask_bias = inf * (mask.astype(jnp.float32) - 1.0)
 
     triangle_bias = _linear(x, params["linear"]["kernel"], precision)
     triangle_bias = jnp.transpose(triangle_bias, (0, 3, 1, 2))
